* Fixed: Wrong ushort vs unsigned in Windows drivers. Patch from Grzegorz (gradha).

+++ tvision/include/tv/win32/screen.h
@@ -69,7 +69,7 @@ protected:
  // Default: static void   setVideoModeExt(char *mode);
  // Default: static void   getCharacters(unsigned offset,ushort *buf,unsigned count);
  // Default: static ushort getCharacter(unsigned dst);
- static void   setCharacter(unsigned offset, ushort value);
+ static void   setCharacter(unsigned offset, unsigned int value);
  static void   setCharacters(unsigned dst, ushort *src, unsigned len);
  static int    System(const char *command, pid_t *pidChild, int in, int out,
                       int err);
